[CookieStoreAPI] Disable API for microsoft.com with quirk until CFNetwork event listener issue is resolved https://bugs.webkit.org/show_bug.cgi?id=286705 rdar://142709866
Reviewed by NOBODY (OOPS!).

CFNetwork is firing an event even when deleting a cookie that doesn't exist.
Microsoft's event listener, upon firing, is deleting a cookie that doesn't exist.
This is creating an infinite loop that is using 100& of the CPU. CFNetwork is
working on a fix and so we're temporarily adding a quirk to disable the API
for the site until that fix lands.

* Source/WebCore/Modules/cookie-store/DOMWindow+CookieStore.idl:
* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::shouldDisableCookieStoreQuirk const):
(WebCore::handleMicrosoftQuirks):
(WebCore::Quirks::determineRelevantQuirks):
* Source/WebCore/page/Quirks.h:
* Source/WebCore/page/QuirksData.h: