Permissions API reports the state of the Geolocation permission as "prompt" in cases where it should report "denied" https://bugs.webkit.org/show_bug.cgi?id=289664 rdar://146917202
Reviewed by NOBODY (OOPS!).

Consider the following scenario:

1. The user has denied permission for a certain API to be used.
2. A site asks for the permission state of this API without ever having made a request
   for permission to use that API.

In this case, as a fingerprinting countermeasure, the Permissions API will reply that the
permission state is "prompt" rather than "denied". We do this to reduce the amount of
fingerprinting information that sites may collect from APIs that they don't intend to use
for actual user-facing functionality.

If a site truly wants to use an API and needs to know the true permission state to do that,
it can request permission to use that API. WebKit will document that this request was made
and any subsequent requests from that site to query the permission state will be given a
truthful response.

We first did this when adding support to the Permissions API for the state of the Notifications
API back in https://commits.webkit.org/252969@main.

The default state is to turn all "denied" states into "prompt" unless we have a record of
an origin requesting permission to use the site. WebKit does not currently record these requests
for the Geolocation API, so it returns "prompt" in all cases where the true result is "denied".

This patch adds support for making a note of which origins requested use of the Geolocation API so
that the true "denied" result is returned when appropriate.

* LayoutTests/fast/dom/Geolocation/geolocation-permissions-query-expected.txt: Added.
* LayoutTests/fast/dom/Geolocation/geolocation-permissions-query-fingerprinting-expected.txt: Added.
* LayoutTests/fast/dom/Geolocation/geolocation-permissions-query-fingerprinting.html: Added.
* LayoutTests/fast/dom/Geolocation/geolocation-permissions-query.html: Added.
* LayoutTests/platform/mac-wk1/TestExpectations:

These layout tests verify the both the normal and fingerprinting-avoiding behavior.

* Source/WebKit/UIProcess/API/C/WKPage.cpp:
(WKPageClearGeolocationPermissionState):
* Source/WebKit/UIProcess/API/C/WKPage.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::requestGeolocationPermissionForFrame):
(WebKit::WebPageProxy::queryPermission):

The set of requesting origins is added to upon request and checked upon query.

(WebKit::WebPageProxy::clearGeolocationPermissionState):
* Source/WebKit/UIProcess/WebPageProxy.h:
* Source/WebKit/UIProcess/WebPageProxyInternals.h:

The set of requesting origins is kept in WebPageProxyInternals.

* Tools/WebKitTestRunner/TestController.cpp:
(WTR::TestController::resetStateToConsistentValues):

The set of requesting origns is reset between layout tests.