[OOM] REGRESSION (Glow): StabilityTracer: com.apple.WebKit.Networking at com.apple.CFNetwork: -[NSHTTPCookieStorage _initWithCFHTTPCookieStorage:] https://bugs.webkit.org/show_bug.cgi?id=290472 rdar://130167032
Reviewed by NOBODY (OOPS!).

It is unclear what exactly is causing the crash. But what we can see is:

1. "~NetworkConnectionToWebProcess" destructor is called, which calls:
2. "stopListeningForCookieChangeNotifications" which calls:
3. "nsCookieStorage()" to get the cookie storage object so that it can
   update it's list of subscribed domains.

If a cookie storage object does not exist, "nsCookieStorage()" will make
one using "_initWithCFHTTPCookieStorage". The crash follows after this point.

But if a cookie storage object doesn't exist, we shouldn't need to update
a non-existent list of subscribed domains. So we add a "nsCookieStorageIfExists()"
function to avoid this creation altogether.

This is a speculative fix.

* Source/WebCore/platform/network/NetworkStorageSession.h:
* Source/WebCore/platform/network/cocoa/NetworkStorageSessionCocoa.mm:
(WebCore::NetworkStorageSession::nsCookieStorageIfExists const):
(WebCore::NetworkStorageSession::nsCookieStorage const):
(WebCore::NetworkStorageSession::stopListeningForCookieChangeNotifications):