StabilityTracer: MobileNotes at WebKit: WebKit::AuxiliaryProcessProxy::protectedConnection const rdar://145887512 https://bugs.webkit.org/show_bug.cgi?id=290763
Reviewed by NOBODY (OOPS!).

WebPageProxy::willPerformPasteCommand() queues an async IPC to
the Network Process. It's ok if the Network Process hasn't launched
yet because the IPC is queued.

But using waitForAsyncReplyAndDispatchImmediately requires the
Network Process to be fully launched. We're hitting the release
assert in AuxiliaryProcessProxy::protectedConnection because
the Network Process hasn't fully launched at the time of use.

But we don't need to hang the UI Process here until it recievies
a response from the Network Process because there are no synchronous
operations happening right after it. So we can simply remove the
call to waitForAsyncReplyAndDispatchImmediately.

So now WebPageProxy sends an IPC to the Network Process to grant
the Web Process file access for that file URL and the Network Process
will send IPC to the Web Process once access is granted.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::executeEditCommand):