[Navigation API] navigation-traverseTo-in-iframe-same-document-preventDefault.html is failing https://bugs.webkit.org/show_bug.cgi?id=297136 rdar://157877259
Reviewed by NOBODY (OOPS!).

According to the layout test, a subframe's navigation should be canceled
if a parent frame has preventDefault set.

This was working prior to https://commits.webkit.org/298121@main.

HistoryController::goToItemForNavigationAPI() recursively collects the
frames that are navigating. If any of them have preventDefault set, it
stops any other navigations from happening.

The reason it was working before 298121@main was because this recursive
collection would always start at the main frame. Even in cases where the
navigation was triggered by an iframe. So the main frame would end up on
the list of navigating frames, we sould see it has preventDefault set,
and we wouldn't allow the iframe navigation to happen.

298121@main changed this so that the recusive collection always starts at
the triggering frame. So in the case of an iframe, the main frame won't
end up in that list. This now means that in the case of an iframe-initiated
navigation, we don't check if the main frame has preventDefault set. So
even if it does, the iframe's navigation is allowed to continue.

We fix this by explicitly checking if any parent frames have preventDefault
set and stopping the triggering frame's navigation if so.

Thank you to Basuke Suzuki for helping come up with the fix :).

* LayoutTests/platform/glib/TestExpectations:
* LayoutTests/platform/ios/TestExpectations:
* LayoutTests/platform/mac-wk2/TestExpectations:
* Source/WebCore/loader/HistoryController.cpp:
(WebCore::frameHasPreventDefault):
(WebCore::HistoryController::goToItemForNavigationAPI):
* Source/WebCore/page/FrameTree.cpp:
(WebCore:: const):
* Source/WebCore/page/FrameTree.h:
* Source/WebCore/page/Navigation.h: